[LeetCode Sync] Runtime - 8 ms (67.53%), Memory - 48.2 MB (24.82%)

+++ my-folder/0907-koko-eating-bananas/solution.java
@@ -66,7 +66,7 @@ class Solution {
         //Iterate through all the piles
         //At by speed add up the time taken to finish the full piles
         //Return true if total number of hrs <= h
-        long totalHrs=0;
+        int totalHrs=0;
         for(long pile:piles){
             totalHrs+=(pile-1)/k+1;
         }
